new Update commited

+++ Dentist-ms/src/main/java/com/example/dentistms/Repository/DentistRepo.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface DentistRepo extends JpaRepository<Dentist, Integer> {
+public interface DentistRepo extends JpaRepository<Dentist, Long> {
     Optional<Dentist> findByEmail(String email);
 }
